+++ test/cases/option-renderStage-default/webpack.config.js
@@ -1,14 +1,13 @@
 const path = require('path');
-const HtmlBundlerPlugin = require('html-bundler-webpack-plugin');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
+const CompressionPlugin = require('compression-webpack-plugin');
+const Compilation = require('webpack/lib/Compilation');
 
 module.exports = {
   mode: 'production',
-  //stats: 'normal',
-  stats: 'errors-warnings',
 
   output: {
     path: path.join(__dirname, 'dist/'),
-    clean: true,
   },
 
   resolve: {
@@ -18,53 +17,37 @@ module.exports = {
   },
 
   plugins: [
+    new CompressionPlugin(),
     new HtmlBundlerPlugin({
       entry: {
-        index: './src/home.html',
-        news: './src/news.html',
+        index: 'src/index.html',
       },
-
       js: {
         filename: 'js/[name].[contenthash:8].js',
       },
-
       css: {
         filename: 'css/[name].[contenthash:8].css',
       },
-
-      //verbose: true,
-      hotUpdate: false, // test to disable auto-injection of hot-update.js files
+      minify: true,
+      // test default value of the renderStage option
+      //renderStage: Infinity + 1, // should throw an error
+      //renderStage: Compilation.PROCESS_ASSETS_STAGE_OPTIMIZE, // the value is too low, it will be set as PROCESS_ASSETS_STAGE_SUMMARIZE, pass OK
     }),
   ],
 
   module: {
     rules: [
       {
-        test: /\.(css|scss)/,
+        test: /\.css$/,
         use: ['css-loader'],
-        //use: ['css-loader', 'sass-loader'],
       },
-
       {
-        test: /\.(png|jpe?g|ico|svg)$/,
+        test: /\.(ico|png|jp?g|svg)/,
         type: 'asset/resource',
         generator: {
           filename: 'img/[name].[hash:8][ext]',
         },
       },
     ],
-  },
-
-  devServer: {
-    //hot: false,
-    static: {
-      directory: path.join(__dirname, 'dist'),
-    },
-    watchFiles: {
-      paths: ['src/**/*.*'],
-      options: {
-        usePolling: true,
-      },
-    },
   },
 };