+++ test/manual/cache-filesystem-rebuild-image-minimizer/src/index.js
@@ -1,3 +1,0 @@
-import img from './img.jpg';
-
-console.log('test', img);