+++ test/manual/cache-filesystem-rebuild-image-minimizer/src/main.js
@@ -1,0 +1,3 @@
+const img = require('./token.svg');
+
+console.log(img);